refactor(popup): group tracked-tab control under src/popup/tabs/ and ratchet the baseline
Moves tab-tracking.ts and tab-tracking-api.ts into src/popup/tabs/, mirroring
the src/lib/tabs/ module they are the popup half of. src/popup goes 14 files
-> 5 and drops out of .folder-size-baseline.json.

Evidence: tab-tracking.ts is the only importer of tab-tracking-api.ts, they
are the strongest co-change pair in the folder over its whole history
(2 focused commits, affinity 0.25 — the folder's history is thin, which is
why the reference graph leads here), and they are the only two popup modules
that reach into src/lib/tabs/ for the shared trackTab/untrackTab helpers
(request-audit, tab-focus, tab-tracking-core, tracked-tab-storage,
cloaked-domains). Nothing else in src/popup touches that surface except
shell/ui-utils.ts, which borrows internal-url for a display check.

What is left at the root of src/popup is not a leftover bucket: ai-web-pilot,
draw-mode, feature-toggles, settings and update-button are five independent
single-purpose controls, each imported exactly once by src/popup.ts, each
owning a different feature, and none sharing a module with any other.

Contract test repointed rather than deleted: tests/extension/
entry-point-parity.test.js reads sources through SRC_ROOT-relative strings
('popup/tab-tracking-api.ts'), invisible to both import rewiring and to a grep
for src/popup/. It now pins popup/tabs/tab-tracking-api.ts and still asserts
the same rule-19 parity invariants against background/ui/context-menus.ts.

Other non-obvious path references updated: popup-tab-tracking-branding,
popup-audit-button and popup-untrack-storage dynamically import the compiled
extension/popup/... modules by string; runtime-log-branding pins the src path
as a literal; docs code_paths in auto-fix, tab-tracking-ux and
lazy-server-start, plus docs/architecture/audit-workflow.md,
docs/architecture/tracked-tab-hover-quick-actions.md and finding F2 in
docs/core/bug-class-audit.md.

8 stale compiled artifacts under extension/popup/ were git rm'd; the rebuilt
output lands in extension/popup/tabs/.
