+++ .folder-size-baseline.json
@@ -25,7 +25,6 @@
     "internal/util": 16,
     "scripts": 13,
     "scripts/docs": 16,
-    "src/popup": 14,
     "src/types": 22
   }
 }
