+++ docs/features/feature/tab-tracking-ux/index.md
@@ -18,8 +18,8 @@ code_paths:
   - src/content/ui/tracked-hover-launcher.ts
   - src/popup.ts
   - src/popup/shell/logo-motion.ts
-  - src/popup/tab-tracking.ts
-  - src/popup/tab-tracking-api.ts
+  - src/popup/tabs/tab-tracking.ts
+  - src/popup/tabs/tab-tracking-api.ts
   - extension/popup.html
   - extension/popup.css
   - src/background/message-handlers.ts
