+++ extension/popup.js
@@ -15,7 +15,7 @@ import { setupRecordingUI } from './popup/recording/recording.js';
 import { setupDrawModeButton } from './popup/draw-mode.js';
 import { setupActionRecordingUI } from './popup/recording/action-recording.js';
 import { FEATURE_TOGGLES as TOGGLE_DEFS, applyFeatureToggles } from './popup/feature-toggles.js';
-import { initTrackPageButton } from './popup/tab-tracking.js';
+import { initTrackPageButton } from './popup/tabs/tab-tracking.js';
 import { applyAiWebPilotToggle } from './popup/ai-web-pilot.js';
 import { initPopupLogoMotion } from './popup/shell/logo-motion.js';
 import { applyWebSocketMode, handleWebSocketModeChange, handleClearLogs, resetClearConfirm } from './popup/settings.js';
@@ -25,7 +25,7 @@ export { updateConnectionStatus };
 export { FEATURE_TOGGLES, initFeatureToggles, applyFeatureToggles } from './popup/feature-toggles.js';
 export { handleFeatureToggle } from './popup/feature-toggles.js';
 export { initAiWebPilotToggle, handleAiWebPilotToggle, applyAiWebPilotToggle } from './popup/ai-web-pilot.js';
-export { initTrackPageButton, handleTrackPageClick } from './popup/tab-tracking.js';
+export { initTrackPageButton, handleTrackPageClick } from './popup/tabs/tab-tracking.js';
 export { handleWebSocketModeChange } from './popup/settings.js';
 export { initWebSocketModeSelector, applyWebSocketMode } from './popup/settings.js';
 export { isInternalUrl } from './popup/shell/ui-utils.js';
