+++ extension/popup/tabs/tab-tracking-api.d.ts
