+++ extension/popup/tabs/tab-tracking-api.js
@@ -3,13 +3,13 @@
  * Why: Separates browser API side-effects from DOM UI state rendering in tab-tracking.
  * Docs: docs/features/feature/tab-tracking-ux/index.md
  */
-import { KABOOM_LOG_PREFIX } from '../lib/brand.js';
-import { StorageKey } from '../lib/constants.js';
-import { getLocal, persist } from '../lib/storage-utils.js';
-import { clearTrackedTab } from '../lib/tabs/tracked-tab-storage.js';
-import { trackTab, untrackTab } from '../lib/tabs/tab-tracking-core.js';
-import { focusTabAndWindow } from '../lib/tabs/tab-focus.js';
-import { requestAudit } from '../lib/tabs/request-audit.js';
+import { KABOOM_LOG_PREFIX } from '../../lib/brand.js';
+import { StorageKey } from '../../lib/constants.js';
+import { getLocal, persist } from '../../lib/storage-utils.js';
+import { clearTrackedTab } from '../../lib/tabs/tracked-tab-storage.js';
+import { trackTab, untrackTab } from '../../lib/tabs/tab-tracking-core.js';
+import { focusTabAndWindow } from '../../lib/tabs/tab-focus.js';
+import { requestAudit } from '../../lib/tabs/request-audit.js';
 /**
  * Handle launching the tracked-site audit workflow from popup controls.
  */
