+++ extension/popup/tabs/tab-tracking.d.ts
