+++ extension/popup/tabs/tab-tracking.js
@@ -7,10 +7,10 @@
  * @fileoverview Tab Tracking Module for Popup
  * Manages the "Track This Tab" button and tracking status
  */
-import { isInternalUrl } from './shell/ui-utils.js';
-import { StorageKey } from '../lib/constants.js';
-import { getLocals, onStorageChanged } from '../lib/storage-utils.js'; // async API only
-import { isDomainCloaked } from '../lib/tabs/cloaked-domains.js';
+import { isInternalUrl } from '../shell/ui-utils.js';
+import { StorageKey } from '../../lib/constants.js';
+import { getLocals, onStorageChanged } from '../../lib/storage-utils.js'; // async API only
+import { isDomainCloaked } from '../../lib/tabs/cloaked-domains.js';
 import { handleAuditClick, handleStopTracking, handleUrlClick, handleTrackPageClick as handleTrackPageClickAPI } from './tab-tracking-api.js';
 let trackingStorageSyncInstalled = false;
 /**
