+++ src/popup/tabs/tab-tracking-api.ts
@@ -4,13 +4,13 @@
  * Docs: docs/features/feature/tab-tracking-ux/index.md
  */
 
-import { KABOOM_LOG_PREFIX } from '../lib/brand.js'
-import { StorageKey } from '../lib/constants.js'
-import { getLocal, persist } from '../lib/storage-utils.js'
-import { clearTrackedTab } from '../lib/tabs/tracked-tab-storage.js'
-import { trackTab, untrackTab } from '../lib/tabs/tab-tracking-core.js'
-import { focusTabAndWindow } from '../lib/tabs/tab-focus.js'
-import { requestAudit } from '../lib/tabs/request-audit.js'
+import { KABOOM_LOG_PREFIX } from '../../lib/brand.js'
+import { StorageKey } from '../../lib/constants.js'
+import { getLocal, persist } from '../../lib/storage-utils.js'
+import { clearTrackedTab } from '../../lib/tabs/tracked-tab-storage.js'
+import { trackTab, untrackTab } from '../../lib/tabs/tab-tracking-core.js'
+import { focusTabAndWindow } from '../../lib/tabs/tab-focus.js'
+import { requestAudit } from '../../lib/tabs/request-audit.js'
 
 export type ShowStateFn = (btn: HTMLButtonElement) => void
 export type ShowTrackingStateFn = (btn: HTMLButtonElement, url: string | undefined, tabId: number | undefined) => void
