+++ src/popup/tabs/tab-tracking.ts
@@ -9,10 +9,10 @@
  * Manages the "Track This Tab" button and tracking status
  */
 
-import { isInternalUrl } from './shell/ui-utils.js'
-import { StorageKey } from '../lib/constants.js'
-import { getLocals, onStorageChanged } from '../lib/storage-utils.js' // async API only
-import { isDomainCloaked } from '../lib/tabs/cloaked-domains.js'
+import { isInternalUrl } from '../shell/ui-utils.js'
+import { StorageKey } from '../../lib/constants.js'
+import { getLocals, onStorageChanged } from '../../lib/storage-utils.js' // async API only
+import { isDomainCloaked } from '../../lib/tabs/cloaked-domains.js'
 import {
   handleAuditClick,
   handleStopTracking,
