+++ tests/extension/popup-tab-tracking-sync.test.js
@@ -118,7 +118,7 @@ describe('popup tab tracking sync', () => {
     const warning = document.getElementById('no-tracking-warning')
     assert.strictEqual(trackingBar.style.display, 'flex')
     // Audit stays hidden while AUDIT_BUTTON_ENABLED is false (see
-    // src/popup/tab-tracking.ts). The tracking bar itself must still appear.
+    // src/popup/tabs/tab-tracking.ts). The tracking bar itself must still appear.
     assert.strictEqual(auditButton.style.display, 'none')
     assert.strictEqual(warning.style.display, 'none')
   })
